+++ public/index.php
@@ -62,6 +62,8 @@ $router->get('/ordenes/{id}/imprimir', [OrdenController::class, 'imprimir']);
 $router->get('/ordenes/{id}/pdf', [OrdenController::class, 'pdf']);
 $router->post('/ordenes/{id}/evidencia', [OrdenController::class, 'subirEvidencia']);
 $router->get('/ordenes/{id}/evidencia/{archivo}', [OrdenController::class, 'verEvidencia']);
+$router->post('/ordenes/{id}/refacciones', [OrdenController::class, 'aplicarRefaccion']);
+$router->post('/ordenes/{id}/refacciones/{uso}/cancelar', [OrdenController::class, 'cancelarRefaccion']);
 
 $router->post('/diagnosticos', [DiagnosticoController::class, 'store']);
 $router->post('/cotizaciones', [CotizacionController::class, 'store']);